+++ include/openpose/core/enumClasses.hpp
@@ -9,7 +9,9 @@ namespace op
         NetOutputResolution,
         OutputResolution,
         ZeroToOne, // [0, 1]
+        ZeroToOneFixedAspect, // [0, 1]
         PlusMinusOne, // [-1, 1]
+        PlusMinusOneFixedAspect, // [-1, 1]
         UnsignedChar, // [0, 255]
         NoScale,
     };