+++ include/openpose/face/faceExtractorCaffe.hpp
@@ -21,7 +21,7 @@ namespace op
         FaceExtractorCaffe(const Point<int>& netInputSize, const Point<int>& netOutputSize,
                            const std::string& modelFolder, const int gpuId,
                            const std::vector<HeatMapType>& heatMapTypes = {},
-                           const ScaleMode heatMapScaleMode = ScaleMode::ZeroToOne,
+                           const ScaleMode heatMapScaleMode = ScaleMode::ZeroToOneFixedAspect,
                            const bool enableGoogleLogging = true);
 
         virtual ~FaceExtractorCaffe();