+++ include/openpose/face/faceExtractorNet.hpp
@@ -20,7 +20,7 @@ namespace op
          */
         explicit FaceExtractorNet(const Point<int>& netInputSize, const Point<int>& netOutputSize,
                                   const std::vector<HeatMapType>& heatMapTypes = {},
-                                  const ScaleMode heatMapScaleMode = ScaleMode::ZeroToOne);
+                                  const ScaleMode heatMapScaleMode = ScaleMode::ZeroToOneFixedAspect);
 
         /**
          * Virtual destructor of the HandExtractor class.