+++ include/openpose/hand/handExtractorCaffe.hpp
@@ -27,7 +27,7 @@ namespace op
                            const std::string& modelFolder, const int gpuId,
                            const int numberScales = 1, const float rangeScales = 0.4f,
                            const std::vector<HeatMapType>& heatMapTypes = {},
-                           const ScaleMode heatMapScaleMode = ScaleMode::ZeroToOne,
+                           const ScaleMode heatMapScaleMode = ScaleMode::ZeroToOneFixedAspect,
                            const bool enableGoogleLogging = true);
 
         /**