+++ include/openpose/pose/poseExtractorCaffe.hpp
@@ -19,7 +19,7 @@ namespace op
         PoseExtractorCaffe(
             const PoseModel poseModel, const std::string& modelFolder, const int gpuId,
             const std::vector<HeatMapType>& heatMapTypes = {},
-            const ScaleMode heatMapScaleMode = ScaleMode::ZeroToOne,
+            const ScaleMode heatMapScaleMode = ScaleMode::ZeroToOneFixedAspect,
             const bool addPartCandidates = false, const bool maximizePositives = false,
             const std::string& protoTxtPath = "", const std::string& caffeModelPath = "",
             const float upsamplingRatio = 0.f, const bool enableNet = true,