+++ src/openpose/pose/poseExtractorNet.cpp
@@ -57,10 +57,13 @@ namespace op
         try
         {
             // Error check
-            if (mHeatMapScaleMode != ScaleMode::ZeroToOne && mHeatMapScaleMode != ScaleMode::PlusMinusOne
+            if (mHeatMapScaleMode != ScaleMode::ZeroToOne
+                && mHeatMapScaleMode != ScaleMode::ZeroToOneFixedAspect
+                && mHeatMapScaleMode != ScaleMode::PlusMinusOne
+                && mHeatMapScaleMode != ScaleMode::PlusMinusOneFixedAspect
                 && mHeatMapScaleMode != ScaleMode::UnsignedChar && mHeatMapScaleMode != ScaleMode::NoScale)
-                error("The ScaleMode heatMapScaleMode must be ZeroToOne, PlusMinusOne, UnsignedChar, or NoScale.",
-                      __LINE__, __FUNCTION__, __FILE__);
+                error("The ScaleMode heatMapScaleMode must be ZeroToOne, ZeroToOneFixedAspect, PlusMinusOne,"
+                    " PlusMinusOneFixedAspect or UnsignedChar.", __LINE__, __FUNCTION__, __FILE__);
 
             // Properties - Init to 0
             for (auto& property : mProperties)
@@ -136,7 +139,8 @@ namespace op
                     if (mHeatMapScaleMode != ScaleMode::NoScale)
                     {
                         // Change from [0,1] to [-1,1]
-                        if (mHeatMapScaleMode == ScaleMode::PlusMinusOne)
+                        if (mHeatMapScaleMode == ScaleMode::PlusMinusOne
+                            || mHeatMapScaleMode == ScaleMode::PlusMinusOneFixedAspect)
                             for (auto i = 0u ; i < volumeBodyParts ; i++)
                                 heatMaps[i] = fastTruncate(heatMaps[i]) * 2.f - 1.f;
                         // [0, 255]
@@ -168,7 +172,8 @@ namespace op
                         if (mHeatMapScaleMode != ScaleMode::NoScale)
                         {
                             // Change from [0,1] to [-1,1]
-                            if (mHeatMapScaleMode == ScaleMode::PlusMinusOne)
+                            if (mHeatMapScaleMode == ScaleMode::PlusMinusOne
+                                || mHeatMapScaleMode == ScaleMode::PlusMinusOneFixedAspect)
                                 for (auto i = 0u ; i < channelOffset ; i++)
                                     heatMapsPtr[i] = fastTruncate(heatMapsPtr[i]) * 2.f - 1.f;
                             // [0, 255]
@@ -205,7 +210,8 @@ namespace op
                     if (mHeatMapScaleMode != ScaleMode::NoScale)
                     {
                         // Change from [-1,1] to [0,1]. Note that PAFs are in [-1,1]
-                        if (mHeatMapScaleMode == ScaleMode::ZeroToOne)
+                        if (mHeatMapScaleMode == ScaleMode::ZeroToOne
+                            || mHeatMapScaleMode == ScaleMode::ZeroToOneFixedAspect)
                             for (auto i = 0u ; i < volumePAFs ; i++)
                                 heatMapsPtr[i] = fastTruncate(heatMapsPtr[i], -1.f) * 0.5f + 0.5f;
                         // [0, 255]